Split optimizer and TFOptimizer
